+++ src/Nancy.ViewEngines.Razor/CSharp/CSharpClrTypeResolver.cs
@@ -1,0 +1,52 @@
+namespace Nancy.ViewEngines.Razor.CSharp
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Web.Razor.Tokenizer.Symbols;
+
+    internal class CSharpClrTypeResolver : ClrTypeResolver<CSharpSymbolType, CSharpSymbol>
+    {        
+        public CSharpClrTypeResolver()
+            :base(CSharpSymbolType.Identifier, CSharpSymbolType.Keyword, CSharpSymbolType.Dot, CSharpSymbolType.WhiteSpace)
+        {            
+        }
+
+        protected override TypeNameParserStep ResolveType()
+        {
+            var identifier = this.PopFullIdentifier();
+
+            var step = new TypeNameParserStep(identifier);
+
+            if (this.symbols.Any() && this.symbols.Peek().Type == CSharpSymbolType.LessThan)
+            {
+                this.symbols.Dequeue();
+
+                while (this.symbols.Peek().Type != CSharpSymbolType.GreaterThan)
+                {
+                    step.GenericArguments.Add(this.ResolveType());
+
+                    while (this.symbols.Peek().Type.Equals(this.WhiteSpace) || this.symbols.Peek().Type == CSharpSymbolType.Comma)
+                    {
+                        this.symbols.Dequeue();
+                    }
+                }
+
+                this.symbols.Dequeue();
+            }
+
+            return step;
+        }
+
+        protected override Type ResolvePrimitiveType(string typeName)
+        {
+            var primitives = new Dictionary<string, Type>
+            {
+                {"string", typeof (String)},
+                {"int", typeof (int)}
+            };
+
+            return (primitives.ContainsKey(typeName) ? primitives[typeName] : null);
+        }
+    }
+}
