+++ src/Nancy.ViewEngines.Razor/Nancy.ViewEngines.Razor.csproj
@@ -113,6 +113,8 @@
     </Compile>
     <Compile Include="AttributeValue.cs" />
     <Compile Include="ClrTypeResolver.cs" />
+    <Compile Include="CodeParserHelper.cs" />
+    <Compile Include="CSharp\CSharpClrTypeResolver.cs" />
     <Compile Include="CSharp\CSharpModelCodeGenerator.cs" />
     <Compile Include="CSharp\CSharpRazorViewRenderer.cs" />
     <Compile Include="DefaultRazorConfiguration.cs" />
@@ -133,6 +135,7 @@
     <Compile Include="RazorViewEngineApplicationStartupRegistrations.cs" />
     <Compile Include="UrlHelpers.cs" />
     <Compile Include="VisualBasic\NancyVisualBasicRazorCodeParser.cs" />
+    <Compile Include="VisualBasic\VisualBasicClrTypeResolver.cs" />
     <Compile Include="VisualBasic\VisualBasicModelCodeGenerator.cs" />
     <Compile Include="VisualBasic\VisualBasicRazorViewRenderer.cs" />
     <Compile Include="EncodedHtmlString.cs" />
